Build: WIP: Refactoring loaders modules.
